requirements.txt
pandas
sklearn
torchvision
numpy
tqdm
matplotlib
reformer_pytorch